ef_drawstrategyimpl OK

+++ include/nw4r/ef/ef_particle.h
@@ -161,7 +161,8 @@ public:
             (mParameter.mTextureReverse >> (layer * NUM_OF_TEX_REVERSE) &
              TEX_REVERSE_ST);
 
-        if (reverse == TEX_REVERSE_S || reverse == TEX_REVERSE_ST) {
+        if (static_cast<u32>(reverse) == TEX_REVERSE_S ||
+            static_cast<u32>(reverse) == TEX_REVERSE_ST) {
             scale -= scale * 2;
         }
 
@@ -178,7 +179,8 @@ public:
             (mParameter.mTextureReverse >> (layer * NUM_OF_TEX_REVERSE) &
              TEX_REVERSE_ST);
 
-        if (reverse == TEX_REVERSE_T || reverse == TEX_REVERSE_ST) {
+        if (static_cast<u32>(reverse) == TEX_REVERSE_T ||
+            static_cast<u32>(reverse) == TEX_REVERSE_ST) {
             scale -= scale * 2;
         }
 
@@ -192,7 +194,8 @@ public:
             (mParameter.mTextureReverse >> (layer * NUM_OF_TEX_REVERSE) &
              TEX_REVERSE_ST);
 
-        if (reverse == TEX_REVERSE_S || reverse == TEX_REVERSE_ST) {
+        if (static_cast<u32>(reverse) == TEX_REVERSE_S ||
+            static_cast<u32>(reverse) == TEX_REVERSE_ST) {
             trans = 1;
         }
 
@@ -209,7 +212,8 @@ public:
             (mParameter.mTextureReverse >> (layer * NUM_OF_TEX_REVERSE) &
              TEX_REVERSE_ST);
 
-        if (reverse == TEX_REVERSE_T || reverse == TEX_REVERSE_ST) {
+        if (static_cast<u32>(reverse) == TEX_REVERSE_T ||
+            static_cast<u32>(reverse) == TEX_REVERSE_ST) {
             trans = 1;
         }
